+++ src/HttPlaceholder.Application.Tests/ScheduledJobs/Queries/GetScheduledJobNamesQueryHandlerFacts.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using HttPlaceholder.Application.Interfaces.HostedServices;
+using HttPlaceholder.Application.ScheduledJobs;
 using HttPlaceholder.Application.ScheduledJobs.Queries.GetScheduledJobNames;
 
 namespace HttPlaceholder.Application.Tests.ScheduledJobs.Queries;
