+++ src/HttPlaceholder.Application/HttPlaceholder.Application.csproj
@@ -8,31 +8,30 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
-        <PackageReference Include="Bogus" Version="34.0.2" />
-        <PackageReference Include="IPAddressRange" Version="5.0.0" />
-        <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="11.1.0" />
-        <PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="5.0.17" />
-        <PackageReference Include="Microsoft.AspNetCore.WebUtilities" Version="2.2.0" />
-        <PackageReference Include="Microsoft.Extensions.Configuration" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Http" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.0" />
-        <PackageReference Include="Microsoft.Extensions.Primitives" Version="7.0.0" />
-        <PackageReference Include="Microsoft.OpenApi.Readers" Version="1.6.4" />
-        <PackageReference Include="SixLabors.ImageSharp" Version="3.0.1" />
-        <PackageReference Include="SixLabors.ImageSharp.Drawing" Version="1.0.0-beta15" />
+        <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1"/>
+        <PackageReference Include="Bogus" Version="34.0.2"/>
+        <PackageReference Include="IPAddressRange" Version="5.0.0"/>
+        <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="11.1.0"/>
+        <PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="5.0.17"/>
+        <PackageReference Include="Microsoft.AspNetCore.WebUtilities" Version="2.2.0"/>
+        <PackageReference Include="Microsoft.Extensions.Configuration" Version="7.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="7.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="7.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Http" Version="7.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Primitives" Version="7.0.0"/>
+        <PackageReference Include="Microsoft.OpenApi.Readers" Version="1.6.4"/>
+        <PackageReference Include="NCrontab" Version="3.3.1"/>
+        <PackageReference Include="SixLabors.ImageSharp" Version="3.0.1"/>
+        <PackageReference Include="SixLabors.ImageSharp.Drawing" Version="1.0.0-beta15"/>
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\HttPlaceholder.Common\HttPlaceholder.Common.csproj" />
-        <ProjectReference Include="..\HttPlaceholder.Domain\HttPlaceholder.Domain.csproj" />
+        <ProjectReference Include="..\HttPlaceholder.Common\HttPlaceholder.Common.csproj"/>
+        <ProjectReference Include="..\HttPlaceholder.Domain\HttPlaceholder.Domain.csproj"/>
 
 
     </ItemGroup>
-
-
 
 
 </Project>
