+++ src/HttPlaceholder.Web.Shared/HttPlaceholder.Web.Shared.csproj
@@ -21,7 +21,6 @@
         <ProjectReference Include="..\HttPlaceholder.WebInfrastructure\HttPlaceholder.WebInfrastructure.csproj" />
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="NCrontab" Version="3.3.1" />
         <PackageReference Include="NSwag.AspNetCore" Version="13.19.0" />
         <PackageReference Include="Serilog" Version="2.12.0" />
         <PackageReference Include="Serilog.AspNetCore" Version="7.0.0" />
